Part III: Move Code Around
