+++ rules_bot.py
@@ -1,43 +1,41 @@
 import configparser
-import html
 import logging
 import os
-import time
-import datetime as dtm
 
-from telegram import ParseMode, MessageEntity, ChatAction, Update, Bot
+from telegram import ParseMode, Bot
 from telegram.error import BadRequest, Unauthorized
 from telegram.ext import (
     CommandHandler,
     Updater,
     MessageHandler,
     Filters,
-    CallbackContext,
     Defaults,
 )
-from telegram.utils.helpers import escape_markdown
 
 from components import inlinequeries, taghints
+from components.callbacks import (
+    start,
+    rules,
+    docs,
+    wiki,
+    help_callback,
+    off_on_topic,
+    sandwich,
+    github,
+    delete_new_chat_members_message,
+    greet_new_chat_members,
+)
 from components.errorhandler import error_handler
-from const import (
-    ENCLOSING_REPLACEMENT_CHARACTER,
-    GITHUB_PATTERN,
-    OFFTOPIC_CHAT_ID,
+from components.const import (
     OFFTOPIC_RULES,
     OFFTOPIC_USERNAME,
     ONTOPIC_RULES,
     ONTOPIC_USERNAME,
-    ONTOPIC_RULES_MESSAGE_LINK,
-    OFFTOPIC_RULES_MESSAGE_LINK,
     ONTOPIC_RULES_MESSAGE_ID,
     OFFTOPIC_RULES_MESSAGE_ID,
 )
-from util import (
-    get_reply_id,
-    reply_or_edit,
-    get_text_not_in_entities,
+from components.util import (
     github_issues,
-    rate_limit,
     rate_limit_tracker,
 )
 
@@ -49,282 +47,9 @@ else:
     logging.basicConfig(
         format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO
     )
-    aps_logger = logging.getLogger('apscheduler')
-    aps_logger.setLevel(logging.WARNING)
+    logging.getLogger('apscheduler').setLevel(logging.WARNING)
 
 logger = logging.getLogger(__name__)
-
-SELF_CHAT_ID = '@'  # Updated in main()
-
-# Welcome new chat members at most ever X minutes
-NEW_CHAT_MEMBERS_LIMIT_SPACING = 60
-
-
-def start(update: Update, context: CallbackContext):
-    args = context.args
-    if args:
-        if args[0] == 'inline-help':
-            inlinequery_help(update, context)
-    elif update.message.chat.username not in (OFFTOPIC_USERNAME, ONTOPIC_USERNAME):
-        update.message.reply_text(
-            "Hi. I'm a bot that will announce the rules of the "
-            "python-telegram-bot groups when you type /rules."
-        )
-
-
-def inlinequery_help(update: Update, context: CallbackContext):
-    chat_id = update.message.chat_id
-    char = ENCLOSING_REPLACEMENT_CHARACTER
-    text = (
-        f"Use the `{char}`-character in your inline queries and I will replace "
-        f"them with a link to the corresponding article from the documentation or wiki.\n\n"
-        f"*Example:*\n"
-        f"{escape_markdown(SELF_CHAT_ID)} I 💙 {char}InlineQueries{char}, "
-        f"but you need an {char}InlineQueryHandler{char} for it.\n\n"
-        f"*becomes:*\n"
-        f"I 💙 [InlineQueries]("
-        "https://python-telegram-bot.readthedocs.io/en/latest/telegram.html#telegram"
-        f".InlineQuery), but you need an [InlineQueryHandler]("
-        f"https://python-telegram-bot.readthedocs.io/en"
-        f"/latest/telegram.ext.html#telegram.ext.InlineQueryHandler) for it.\n\n"
-        f"Some wiki pages have spaces in them. Please replace such spaces with underscores. "
-        f"The bot will automatically change them back desired space."
-    )
-    context.bot.sendMessage(
-        chat_id, text, parse_mode=ParseMode.MARKDOWN, disable_web_page_preview=True
-    )
-
-
-@rate_limit
-def rules(update: Update, _: CallbackContext):
-    """Load and send the appropriate rules based on which group we're in"""
-    if update.message.chat.username == ONTOPIC_USERNAME:
-        update.message.reply_text(ONTOPIC_RULES, disable_web_page_preview=True, quote=False)
-        update.message.delete()
-    elif update.message.chat.username == OFFTOPIC_USERNAME:
-        update.message.reply_text(OFFTOPIC_RULES, disable_web_page_preview=True, quote=False)
-        update.message.delete()
-    else:
-        update.message.reply_text(
-            "Hmm. You're not in a python-telegram-bot group, "
-            "and I don't know the rules around here."
-        )
-
-
-@rate_limit
-def docs(update: Update, _: CallbackContext):
-    """ Documentation link """
-    text = (
-        "You can find our documentation at "
-        "[Read the Docs](https://python-telegram-bot.readthedocs.io/en/stable/)"
-    )
-    if update.message.reply_to_message:
-        reply_id = update.message.reply_to_message.message_id
-    else:
-        reply_id = None
-    update.message.reply_text(
-        text,
-        parse_mode='Markdown',
-        quote=False,
-        disable_web_page_preview=True,
-        reply_to_message_id=reply_id,
-    )
-    update.message.delete()
-
-
-@rate_limit
-def wiki(update: Update, _: CallbackContext):
-    """ Wiki link """
-    text = (
-        "You can find our wiki on "
-        "[GitHub](https://github.com/python-telegram-bot/python-telegram-bot/wiki)"
-    )
-    if update.message.reply_to_message:
-        reply_id = update.message.reply_to_message.message_id
-    else:
-        reply_id = None
-    update.message.reply_text(
-        text,
-        parse_mode='Markdown',
-        quote=False,
-        disable_web_page_preview=True,
-        reply_to_message_id=reply_id,
-    )
-    update.message.delete()
-
-
-@rate_limit
-def help_callback(update: Update, context: CallbackContext):
-    """ Link to rules readme """
-    text = (
-        f'You can find an explanation of @{html.escape(context.bot.username)}\'s functionality '
-        'wiki on <a href="https://github.com/python-telegram-bot/rules-bot/blob/master/README.md">'
-        'GitHub</a>.'
-    )
-    if update.message.reply_to_message:
-        reply_id = update.message.reply_to_message.message_id
-    else:
-        reply_id = None
-    update.message.reply_text(
-        text,
-        quote=False,
-        disable_web_page_preview=True,
-        reply_to_message_id=reply_id,
-    )
-    update.message.delete()
-
-
-def off_on_topic(update: Update, context: CallbackContext):
-    chat_username = update.message.chat.username
-    group_one = context.match.group(1)
-    if chat_username == ONTOPIC_USERNAME and group_one.lower() == 'off':
-        reply = update.message.reply_to_message
-        moved_notification = 'I moved this discussion to the [off-topic Group]({}).'
-        if reply and reply.text:
-            issued_reply = get_reply_id(update)
-
-            if reply.from_user.username:
-                name = '@' + reply.from_user.username
-            else:
-                name = reply.from_user.first_name
-
-            replied_message_text = reply.text_html
-            replied_message_id = reply.message_id
-
-            text = (
-                f'{name} <a href="t.me/pythontelegrambotgroup/{replied_message_id}">wrote</a>:\n'
-                f'{replied_message_text}\n\n'
-                f'⬇️ ᴘʟᴇᴀsᴇ ᴄᴏɴᴛɪɴᴜᴇ ʜᴇʀᴇ ⬇️'
-            )
-
-            offtopic_msg = context.bot.send_message(
-                OFFTOPIC_CHAT_ID, text, disable_web_page_preview=True
-            )
-
-            update.message.reply_text(
-                moved_notification.format(
-                    'https://telegram.me/pythontelegrambottalk/' + str(offtopic_msg.message_id)
-                ),
-                disable_web_page_preview=True,
-                parse_mode=ParseMode.MARKDOWN,
-                reply_to_message_id=issued_reply,
-            )
-
-        else:
-            update.message.reply_text(
-                'The off-topic group is [here](https://telegram.me/pythontelegrambottalk). '
-                'Come join us!',
-                disable_web_page_preview=True,
-                parse_mode=ParseMode.MARKDOWN,
-            )
-
-    elif chat_username == OFFTOPIC_USERNAME and group_one.lower() == 'on':
-        update.message.reply_text(
-            'The on-topic group is [here](https://telegram.me/pythontelegrambotgroup). '
-            'Come join us!',
-            disable_web_page_preview=True,
-            parse_mode=ParseMode.MARKDOWN,
-        )
-
-
-def sandwich(update: Update, context: CallbackContext):
-    if update.message.chat.username == OFFTOPIC_USERNAME:
-        if 'sudo' in context.match.group(0):
-            update.message.reply_text("Okay.", quote=True)
-        else:
-            update.message.reply_text("What? Make it yourself.", quote=True)
-
-
-def keep_typing(last, chat, action):
-    now = time.time()
-    if (now - last) > 1:
-        chat.send_action(action)
-    return now
-
-
-def github(update: Update, context: CallbackContext):
-    message = update.effective_message
-    last = 0
-    thing_matches = []
-    things = {}
-
-    # Due to bug in ptb we need to convert entities of type URL to TEXT_LINK
-    # for them to be converted to html
-    for entity in message.entities:
-        if entity.type == MessageEntity.URL:
-            entity.type = MessageEntity.TEXT_LINK
-            entity.url = message.parse_entity(entity)
-
-    for match in GITHUB_PATTERN.finditer(get_text_not_in_entities(message.text_html)):
-        logging.debug(match.groupdict())
-        owner, repo, number, sha = [
-            match.groupdict()[x] for x in ('owner', 'repo', 'number', 'sha')
-        ]
-        if number or sha:
-            thing_matches.append((owner, repo, number, sha))
-
-    for thing_match in thing_matches:
-        last = keep_typing(last, update.effective_chat, ChatAction.TYPING)
-        owner, repo, number, sha = thing_match
-        if number:
-            issue = github_issues.get_issue(int(number), owner, repo)
-            things[issue.url] = github_issues.pretty_format_issue(issue)
-        elif sha:
-            commit = github_issues.get_commit(sha, owner, repo)
-            things[commit.url] = github_issues.pretty_format_commit(commit)
-
-    if things:
-        reply_or_edit(
-            update,
-            context,
-            '\n'.join([f'<a href="{url}">{name}</a>' for url, name in things.items()]),
-        )
-
-
-def delete_new_chat_members_message(update: Update, _: CallbackContext):
-    update.message.delete()
-
-
-def greet_new_chat_members(update: Update, context: CallbackContext):
-    group_user_name = update.effective_chat.username
-    # Get saved users
-    user_lists = context.chat_data.setdefault('new_chat_members', {})
-    users = user_lists.setdefault(group_user_name, [])
-
-    # save new users
-    new_chat_members = update.message.new_chat_members
-    for user in new_chat_members:
-        users.append(user.mention_html())
-
-    # check rate limit
-    last_message_date = context.chat_data.setdefault(
-        'new_chat_members_timeout',
-        dtm.datetime.now() - dtm.timedelta(minutes=NEW_CHAT_MEMBERS_LIMIT_SPACING + 1),
-    )
-    if dtm.datetime.now() < last_message_date + dtm.timedelta(
-        minutes=NEW_CHAT_MEMBERS_LIMIT_SPACING
-    ):
-        logging.debug('Waiting a bit longer before greeting new members.')
-        return
-
-    # save new timestamp
-    context.chat_data['new_chat_members_timeout'] = dtm.datetime.now()
-
-    link = (
-        ONTOPIC_RULES_MESSAGE_LINK
-        if group_user_name == ONTOPIC_USERNAME
-        else OFFTOPIC_RULES_MESSAGE_LINK
-    )
-    text = (
-        f'Welcome {", ".join(users)}! If you haven\'t already, read the rules of this '
-        f'group and be sure to follow them. You can find them <a href="{link}">here 🔗</a>.'
-    )
-
-    # Clear users list
-    users.clear()
-
-    # send message
-    update.message.reply_text(text, disable_web_page_preview=True, quote=False)
 
 
 def update_rules_messages(bot: Bot):
@@ -357,55 +82,52 @@ def main():
     dispatcher = updater.dispatcher
     update_rules_messages(updater.bot)
 
-    global SELF_CHAT_ID  # pylint: disable=W0603
-    SELF_CHAT_ID = f'@{updater.bot.get_me().username}'
+    dispatcher.add_handler(MessageHandler(~Filters.command, rate_limit_tracker), group=-1)
 
-    rate_limit_tracker_handler = MessageHandler(~Filters.command, rate_limit_tracker)
+    # Note: Order matters!
+    # Taghints - works with regex
+    taghints.register(dispatcher)
 
-    start_handler = CommandHandler('start', start)
-    rules_handler = CommandHandler('rules', rules)
-    rules_handler_hashtag = MessageHandler(Filters.regex(r'.*#rules.*'), rules)
-    docs_handler = CommandHandler('docs', docs)
-    wiki_handler = CommandHandler('wiki', wiki)
-    help_handler = CommandHandler('help', help_callback)
-    sandwich_handler = MessageHandler(
-        Filters.regex(r'(?i)[\s\S]*?((sudo )?make me a sandwich)[\s\S]*?'), sandwich
-    )
-    off_on_topic_handler = MessageHandler(
-        Filters.regex(r'(?i)[\s\S]*?\b(?<!["\\])(off|on)[- _]?topic\b'), off_on_topic
-    )
-    delete_new_chat_members_handler = MessageHandler(
-        Filters.status_update.new_chat_members, delete_new_chat_members_message
-    )
-    greet_new_chat_members_handler = MessageHandler(
-        Filters.status_update.new_chat_members, greet_new_chat_members
-    )
+    # Simple commands
+    dispatcher.add_handler(CommandHandler('start', start))
+    dispatcher.add_handler(CommandHandler('rules', rules))
+    dispatcher.add_handler(MessageHandler(Filters.regex(r'.*#rules.*'), rules))
+    dispatcher.add_handler(CommandHandler('docs', docs))
+    dispatcher.add_handler(CommandHandler('wiki', wiki))
+    dispatcher.add_handler(CommandHandler('help', help_callback))
 
+    # Stuff that runs on every message with regex
+    dispatcher.add_handler(
+        MessageHandler(
+            Filters.regex(r'(?i)[\s\S]*?((sudo )?make me a sandwich)[\s\S]*?'), sandwich
+        )
+    )
+    dispatcher.add_handler(
+        MessageHandler(
+            Filters.regex(r'(?i)[\s\S]*?\b(?<!["\\])(off|on)[- _]?topic\b'), off_on_topic
+        )
+    )
     # We need several matches so Filters.regex is basically useless
     # therefore we catch everything and do regex ourselves
     # This should probably be in another dispatcher group
-    # but I kept getting SystemErrors...
-    github_handler = MessageHandler(
-        Filters.text & Filters.update.messages & ~Filters.command, github
+    # but I kept getting SystemErrors..
+    dispatcher.add_handler(
+        MessageHandler(Filters.text & Filters.update.messages & ~Filters.command, github)
     )
 
-    dispatcher.add_handler(rate_limit_tracker_handler, group=-1)
+    # Status updates
+    dispatcher.add_handler(
+        MessageHandler(Filters.status_update.new_chat_members, greet_new_chat_members)
+    )
+    dispatcher.add_handler(
+        MessageHandler(Filters.status_update.new_chat_members, delete_new_chat_members_message),
+        group=1,
+    )
 
-    # Note: Order matters!
-    taghints.register(dispatcher)
-    dispatcher.add_handler(start_handler)
-    dispatcher.add_handler(rules_handler)
-    dispatcher.add_handler(rules_handler_hashtag)
-    dispatcher.add_handler(docs_handler)
-    dispatcher.add_handler(wiki_handler)
-    dispatcher.add_handler(help_handler)
-    dispatcher.add_handler(sandwich_handler)
-    dispatcher.add_handler(off_on_topic_handler)
-    dispatcher.add_handler(github_handler)
-    dispatcher.add_handler(greet_new_chat_members_handler)
-    dispatcher.add_handler(delete_new_chat_members_handler, group=1)
-
+    # Inline Queries
     inlinequeries.register(dispatcher)
+
+    # Error Handler
     dispatcher.add_error_handler(error_handler)
 
     updater.start_polling()
